requirements.txt
# LLM APIs
anthropic
backoff
openai
# Viz
matplotlib
pypdf
pymupdf4llm
seaborn
# Common Requirements
numpy
transformers
datasets
tiktoken
wandb
tqdm
rich
humanize
dataclasses-json
funcy
black
genson
shutup
python-igraph
coolname
jsonschema
omegaconf
botocore
boto3
sentence_transformers
accelerate
Pillow